+++ src/components/landing_page/ChooseHunt.tsx
@@ -17,8 +17,8 @@ import { styled } from "@mui/material/styles";
 import MuiToggleButton from "@mui/material/ToggleButton";
 import path from "path";
 import React, { ChangeEvent, useEffect, useState } from "react";
-import { ExitableModal } from "src/components/ExitableModal";
-import { PageHeaderAndSubtitle } from "src/components/PageHeaderAndSubtitle";
+import { ExitableModal } from "src/components/reusable/ExitableModal";
+import { PageHeaderAndSubtitle } from "src/components/reusable/PageHeaderAndSubtitle";
 import { logger } from "src/logger";
 import { resetStorage } from "src/providers/helpers";
 import { getLastError, getURL } from "src/providers/runtime";
