+++ src/components/landing_page/CreateYourOwn.tsx
@@ -1,6 +1,6 @@
 import { CardMedia, Container, Grid, Link } from "@mui/material";
 import React from "react";
-import { PageHeaderAndSubtitle } from "src/components/PageHeaderAndSubtitle";
+import { PageHeaderAndSubtitle } from "src/components/reusable/PageHeaderAndSubtitle";
 
 export const CreateYourOwn = () => (
   <>
