+++ src/components/landing_page/Tutorial.tsx
@@ -1,6 +1,6 @@
 import { Container, Grid, Typography } from "@mui/material";
 import React from "react";
-import { PageHeaderAndSubtitle } from "src/components/PageHeaderAndSubtitle";
+import { PageHeaderAndSubtitle } from "src/components/reusable/PageHeaderAndSubtitle";
 
 export const Tutorial = () => (
   <>
