+++ src/components/reusable/Header.tsx
@@ -1,6 +1,6 @@
 import { Box } from "@mui/material";
 import React from "react";
-import { PageHeaderAndSubtitle } from "src/components/PageHeaderAndSubtitle";
+import { PageHeaderAndSubtitle } from "src/components/reusable/PageHeaderAndSubtitle";
 
 export const Header = () => (
   <>
